+++ src/components/ProjectView.vue
@@ -44,7 +44,7 @@
               <thead>
               <tr>
                 <th class="status-table-header" scope="col">Data Request Number (DRN)</th>
-                <th v-if="visibleBridgeheads?.length == 1" class="status-table-header" scope="col">Votum</th>
+                <th v-if="visibleBridgeheads?.length == 1 && project?.state !== 'DRAFT'" class="status-table-header" scope="col">Votum</th>
                 <th v-if="visibleBridgeheads?.length == 1" class="status-table-header" scope="col">Teiler</th>
                 <th v-if="visibleBridgeheads?.length == 1" class="status-table-header" scope="col">User Access</th>
                 <th class="status-table-header" v-if="visibleBridgeheads?.length == 1 && dataShieldStatus" scope="col">DataSHIELD Status</th>
@@ -62,7 +62,7 @@
               <tbody>
               <tr>
                 <td>{{ project ? project.code : '' }}</td>
-                <td v-if="visibleBridgeheads?.length == 1">
+                <td v-if="visibleBridgeheads?.length == 1 && project?.state !== 'DRAFT'">
                   <div>
                     <div v-if="existsVotum" class="states-circle-container">
                       <div class="state_circle green"></div>
@@ -942,7 +942,7 @@ export default defineComponent({
           uploadAction: this.Action.UPLOAD_VOTUM_ACTION,
           downloadAction: this.Action.DOWNLOAD_VOTUM_ACTION,
           downloadModule: this.Module.PROJECT_DOCUMENTS_MODULE,
-          visibilityCondition: !this.existsDraftDialog || this.draftDialogStepper.currentStep === DialogStep.SUMMARY
+          visibilityCondition: this.project?.state !== 'DRAFT' && (!this.existsDraftDialog || this.draftDialogStepper.currentStep === DialogStep.SUMMARY)
         },
         {
           fieldKey: "Votum for all bridgeheads",
@@ -952,7 +952,7 @@ export default defineComponent({
           uploadAction: this.Action.UPLOAD_VOTUM_FOR_ALL_BRIDGEHEADS_ACTION,
           downloadAction: this.Action.DOWNLOAD_VOTUM_FOR_ALL_BRIDGEHEADS_ACTION,
           downloadModule: this.Module.PROJECT_DOCUMENTS_MODULE,
-          visibilityCondition: !this.existsDraftDialog || this.draftDialogStepper.currentStep === DialogStep.SUMMARY
+          visibilityCondition: this.project?.state !== 'DRAFT' && (!this.existsDraftDialog || this.draftDialogStepper.currentStep === DialogStep.SUMMARY)
         },
         {
           fieldKey: "Script",